requirements.txt
Flask
Pillow
stegano
